pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fallacy-detector-ai"
version = "1.0.0"
description = "An AI agent that detects logical fallacies in news articles based on Aristotelian logic"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
requires-python = ">=3.8"
dependencies = [
    "langchain>=0.1.0",
    "langchain-openai>=0.1.0", 
    "langchain-community>=0.1.0",
    "openai>=1.0.0",
    "pandas>=1.5.0",
    "python-dotenv>=1.0.0",
    "beautifulsoup4>=4.11.0",
    "lxml>=4.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/fallacy-detector-ai"
Repository = "https://github.com/yourusername/fallacy-detector-ai.git"

[project.scripts]
fallacy-detector = "fallacy_detector.__main__:main"
fallacy-demo = "fallacy_detector.__main__:demo"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
